feat(ar): run stereo probe in an isolated ":probe" process
The earlier worker-thread probe couldn't actually prevent an ANR: ARCore's
broken-stereo thrash is in native VIO threads that peg the CPU process-wide
and starve the main thread regardless of which Kotlin thread drives update().

Move the probe into StereoProbeService, declared android:process=":probe".
The forced-stereo session and its native thread storm now live in a throwaway
background process; the OS schedules the foreground UI process ahead of it, so
a hang can never ANR the app — and if the probe process is killed, the binding
drops and the client falls back to mono. The verdict comes back over a
Messenger; ArViewModel binds the service, awaits the reply with a timeout
(probe timeout + headroom), caches the tri-state result, and only selects the
stereo camera config when the probe confirmed the device can track.
